requirements.txt
claude-agent-sdk>=0.1.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.7.0
aiosqlite>=0.20.0
rich>=13.7.0
structlog>=24.1.0
numpy>=1.26.0
scipy>=1.13.0
aiohttp>=3.9.0
python-dateutil>=2.9.0
openai>=1.0.0
